[apex] Move AvoidHardCodingId Rule to errorprone
