+++ pmd-apex/src/main/java/net/sourceforge/pmd/lang/apex/rule/errorprone/AvoidHardcodingIdRule.java
@@ -2,7 +2,7 @@
  * BSD-style license; for more info see http://pmd.sourceforge.net/license.html
  */
 
-package net.sourceforge.pmd.lang.apex.rule.bestpractices;
+package net.sourceforge.pmd.lang.apex.rule.errorprone;
 
 import java.util.regex.Pattern;
