+++ pmd-apex/src/main/resources/category/apex/bestpractices.xml
@@ -79,33 +79,6 @@ global class Unchangeable {
         </example>
     </rule>
 
-    <rule name="AvoidHardcodingId"
-          since="6.0.0"
-          message="Hardcoding Id's is bound to break when changing environments."
-          class="net.sourceforge.pmd.lang.apex.rule.bestpractices.AvoidHardcodingIdRule"
-          externalInfoUrl="${pmd.website.baseurl}/pmd_rules_apex_bestpractices.html#avoidhardcodingid">
-        <description>
-When deploying Apex code between sandbox and production environments, or installing Force.com AppExchange packages,
-it is essential to avoid hardcoding IDs in the Apex code. By doing so, if the record IDs change between environments,
-the logic can dynamically identify the proper data to operate against and not fail.
-        </description>
-        <priority>3</priority>
-        <example>
-<![CDATA[
-public without sharing class Foo {
-    void foo() {
-        //Error - hardcoded the record type id
-        if(a.RecordTypeId == '012500000009WAr'){
-            //do some logic here.....
-        } else if(a.RecordTypeId == '0123000000095Km'){
-            //do some logic here for a different record type...
-        }
-    }
-}
-]]>
-        </example>
-    </rule>
-
     <rule name="AvoidLogicInTrigger"
           since="5.5.0"
           message="Avoid logic in triggers"
