+++ pmd-apex/src/main/resources/rulesets/apex/ruleset.xml
@@ -158,7 +158,7 @@
          <property name="cc_block_highlighting" value="false" />
       </properties>
    </rule>
-   <rule ref="category/apex/bestpractices.xml/AvoidHardcodingId" message="Avoid hardcoding ID's">
+   <rule ref="category/apex/errorprone.xml/AvoidHardcodingId" message="Avoid hardcoding ID's">
       <priority>3</priority>
       <properties>
          <!-- relevant for Code Climate output only -->
