+++ pmd-apex/src/main/resources/rulesets/apex/style.xml
@@ -13,10 +13,10 @@ The Style Ruleset contains rules regarding preferred usage of names and identifi
     <rule ref="category/apex/codestyle.xml/VariableNamingConventions" deprecated="true" />
 
     <rule ref="category/apex/errorprone.xml/AvoidDirectAccessTriggerMap" deprecated="true" />
+    <rule ref="category/apex/errorprone.xml/AvoidHardcodingId" deprecated="true" />
     <rule ref="category/apex/errorprone.xml/MethodWithSameNameAsEnclosingClass" deprecated="true" />
 
     <rule ref="category/apex/bestpractices.xml/AvoidGlobalModifier" deprecated="true" />
-    <rule ref="category/apex/bestpractices.xml/AvoidHardcodingId" deprecated="true" />
     <rule ref="category/apex/bestpractices.xml/AvoidLogicInTrigger" deprecated="true" />
 
 </ruleset>
