+++ pmd-core/src/main/resources/rulesets/releases/600.xml
@@ -12,9 +12,9 @@ This ruleset contains links to rules that are new in PMD v6.0.0
     <rule ref="rulesets/java/migrating.xml/ForLoopCanBeForeach"/>
     <rule ref="rulesets/java/strictexception.xml/DoNotExtendJavaLangThrowable"/>
 
-    <rule ref="category/apex/bestpractices.xml/AvoidHardcodingId" />
     <rule ref="category/apex/design.xml/CyclomaticComplexity" />
     <rule ref="category/apex/errorprone.xml/AvoidDirectAccessTriggerMap"/>
+    <rule ref="category/apex/errorprone.xml/AvoidHardcodingId" />
     <rule ref="category/apex/errorprone.xml/EmptyCatchBlock"/>
     <rule ref="category/apex/errorprone.xml/EmptyIfStmt"/>
     <rule ref="category/apex/errorprone.xml/EmptyStatementBlock"/>
